# apps/coupon-api-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DDADAN Coupon API",
    "description": "상점별 쿠폰·도장 발급 시스템 API. 모든 경로는 `/api/coupon/v1` 아래에 있으며, JSON 필드는 snake_case, 식별자는 UUID 문자열, 시각은 RFC 3339 UTC, 금액은 원 단위 정수입니다.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7810",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/coupon/v1/owner/catalog/categories": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "My catalogue categories.",
        "operationId": "list_categories",
        "responses": {
          "200": {
            "description": "Categories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogCategoriesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      },
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Add a category.",
        "operationId": "create_category",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Category created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogCategory"
                }
              }
            }
          },
          "409": {
            "description": "The name is already used in this store"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
    "/api/coupon/v1/owner/catalog/categories/{category_id}": {
      "patch": {
        "tags": [
          "catalog"
        ],
        "summary": "Edit a category.",
        "operationId": "patch_category",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "description": "Category id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "Expected version",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated category",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogCategory"
                }
              }
            }
          },
          "404": {
            "description": "No such category in this store"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
    "/api/coupon/v1/owner/catalog/items": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "My catalogue items.",
        "operationId": "list_items",
        "parameters": [
          {
            "name": "include_inactive",
            "in": "query",
            "description": "Include deactivated items. Off by default, because the common case is picking an\nitem for a new policy, where only active items may be chosen (§8.3).",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Catalogue items",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      },
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Add a catalogue item.",
        "operationId": "create_item",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Item created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            }
          },
          "409": {
            "description": "The SKU is already in use"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
    "/api/coupon/v1/owner/catalog/items/{item_id}": {
      "patch": {
        "tags": [
          "catalog"
        ],
        "summary": "Edit a catalogue item, including deactivating it.",
        "operationId": "patch_item",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "description": "Catalogue item id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "Expected version",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            }
          },
          "404": {
            "description": "No such item in this store"
          },
          "409": {
            "description": "Someone else changed it first"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CatalogCategoriesResponse": {
        "type": "object",
        "required": [
          "categories"
        ],
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogCategory"
            }
          }
        }
      },
      "CatalogCategory": {
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "sort_order",
          "created_at",
          "updated_at",
          "version"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "sort_order": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/ResourceStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CatalogItem": {
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "created_at",
          "updated_at",
          "version"
        ],
        "properties": {
          "category_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "category_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "reference_price": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Reference price only. The price that counts is the one entered at the till (§8.3)."
          },
          "sku": {
            "type": [
              "string",
              "null"
            ],
            "description": "The store's own SKU. Optional — most small shops do not have one (§8.3)."
          },
          "status": {
            "$ref": "#/components/schemas/ResourceStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CatalogItemsResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogItem"
            }
          }
        }
      },
...
      "CreateCategoryRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "sort_order": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
      "CreateItemRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "category_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "reference_price": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "sku": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ResourceStatus": {
        "type": "string",
        "description": "`coupon.resource_status`.",
        "enum": [
          "ACTIVE",
          "INACTIVE",
          "ARCHIVED"
        ]
      },
...
      "UpdateCategoryRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "sort_order": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResourceStatus"
              }
            ]
          },
          "version": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "UpdateItemRequest": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "reference_price": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "sku": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResourceStatus",
                "description": "Deactivating keeps the item in every snapshot that already names it (§8.3)."
              }
            ]
          },
          "version": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "firebase": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token"
      }
    }
  },
...
}